refactor: move helper to api test folder
